Clean up filter node
