+++ autonav_ws/src/autonav_filters/CMakeLists.txt
@@ -14,7 +14,9 @@ ament_python_install_package(${PROJECT_NAME})
 
 # Install Python executables
 install(PROGRAMS
-	src/pf.py
+	src/filters.py
+	src/deadrekt.py
+	src/particlefilter.py
 	DESTINATION lib/${PROJECT_NAME}
 	)
